requirements.txt
Django==4.0.6
django-browser-reload==1.6.0
django-ckeditor==6.4.2
django-environ==0.9.0
django-extensions==3.1.5
django-filter==22.1
django-import-export==2.8.0
django-js-asset==2.0.0
django-tailwind==3.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0